refactor(web): update import paths to use absolute imports and add mock data for charts and members
