+++ apps/web/app/layout.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from 'react';
 
-import '../styles/globals.css';
+import '@/styles/globals.css';
 
 type Props = {
 	children: ReactNode;
